Create 5x5_training.py
Train an AI to play 5x5 tic tac toe. Change the parameter "training_rounds" on line 332 to determine number of training rounds to conduct. "modulo" on line 334 can be changed to show the number of rounds on the progress bar. A policy will be created once the training process is complete. The filename of the policy will indicate how many training rounds were conducted for that particular policy.